refactor(src): 将 selected_part_index 改为 size_t 类型
- 将 int 类型的 selected_part_index 改为 size_t 类型，以匹配 parts.size() 的返回类型
- 这个改动可以提高代码的类型安全性和一致性，避免潜在的类型转换问题

+++ src/plugin-main.cpp
@@ -380,7 +380,7 @@ public slots:
         }
 
         // 填充分区下拉框并设置默认值
-        int selected_part_index = 0;
+        size_t selected_part_index = 0;
         for (size_t i = 0; i < parts.size(); i++) {
             partCombo->addItem(parts[i].name, parts[i].id);
             if (parts[i].id == config.part_id) {
